@@ -1,31 +1,32 @@
-const path = require("path");
-const os = require('os');
+import path from 'path'
+import os from 'os'
+import { fileURLToPath } from 'url';
+import HtmlWebpackPlugin from 'html-webpack-plugin'
+import MiniCssExtractPlugin from 'mini-css-extract-plugin'
+import CopyPlugin from 'copy-webpack-plugin'
+import webpack from 'webpack'
 
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
-const TerserJSPlugin = require('terser-webpack-plugin');
-const CssMinimizerPlugin = require('css-minimizer-webpack-plugin');
-const {CleanWebpackPlugin} = require('clean-webpack-plugin');
-const webpack = require('webpack');
-const CopyPlugin = require('copy-webpack-plugin');
+
+import TerserJSPlugin from 'terser-webpack-plugin';
+import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
+import {CleanWebpackPlugin} from 'clean-webpack-plugin';
 
 // process.traceDeprecation = true;
 
 const getLocalExternalIP = () => [].concat(...Object.values(os.networkInterfaces()))
-    .filter(details => details.family === 'IPv4' && !details.internal)
-    .pop().address
+    .filter(details => (details.family === 'IPv4' || details.family === 4) && !details.internal)
+    .pop()?.address
 
-module.exports = (env, argv) => {
+const webConfig = (env, argv) => {
     const devMode = !argv || (argv.mode !== 'production');
-    let addr = getLocalExternalIP() || '0.0.0.0';
+    const addr = getLocalExternalIP() || '0.0.0.0';
+    const dirname = path.dirname(fileURLToPath(import.meta.url));
     return {
 
         entry: {main: "./src/index.js"},
         output: {
-            path: path.resolve(__dirname, "dist"),
-            filename: devMode ? "[name].js" : "[name].[contenthash].js",
-            publicPath: devMode ? "/" : "./dist/"
-            // publicPath: "./dist/"
+            path: path.resolve(dirname, "docs"),
+            filename: devMode ? "[name].js" : "[name].[contenthash].js"
         },
         module: {
             rules: [
@@ -57,7 +58,6 @@ module.exports = (env, argv) => {
                 template: "./src/index.html",
                 minify: false,
                 scriptLoading: 'defer',
-                filename: devMode ? "./index.html" : "../index.html",
                 inject: 'head'
                 // filename: 'index.html'
             }),
@@ -73,11 +73,19 @@ module.exports = (env, argv) => {
             // })]),
             new webpack.DefinePlugin({
                 __USE_SERVICE_WORKERS__: !devMode
+            }),
+            new CopyPlugin({
+                patterns: [
+                    { from: './images', to: './images' },
+                    { from: './sound', to: './sound' },
+                    { from: './src/manifest.json', to: './' },
+                    { from: './src/rules.html', to: './' },
+                    { from: './.well-known', to: './.well-known' },
+                    { from: './github', to: './' }
+                ],
             })
         ],
         devServer: {
-            // contentBase: path.resolve(__dirname, "src"),
-            historyApiFallback: true,
             compress: true,
             port: 8080,
             hot: true,
@@ -88,3 +96,5 @@ module.exports = (env, argv) => {
         }
     }
 };
+
+export default webConfig;
